Filter and search turned into a form
